refactor : util 함수를 utils.js 파일로 이동
- util 함수들을 모아 utils.js 파일로 이동
- HTML에 있는 scripts 태그에 type = modules 추가
