+++ views/index.html
@@ -8,7 +8,8 @@
     <title>kakaopage</title>
   </head>
   <body>
-    <script src="javascripts/main.js"></script>
+    <script type="module" src="javascripts/utils.js"></script>
+    <script type="module" src="javascripts/main.js"></script>
     <div class="header">
       <div class="header__logo">kakaopage</div>
       <div class="header__status">
